+++ src/Decorators/UserAdminDecorator.php
@@ -9,6 +9,8 @@ use Bozboz\Admin\Fields\EmailField;
 use Bozboz\Admin\Fields\HiddenField;
 use Bozboz\Admin\Fields\PasswordField;
 use Bozboz\Admin\Models\User;
+use Bozboz\Permissions\Permission;
+use Illuminate\Support\Facades\Auth;
 
 class UserAdminDecorator extends ModelAdminDecorator
 {
@@ -27,7 +29,12 @@ class UserAdminDecorator extends ModelAdminDecorator
 
 	public function modifyListingQuery(Builder $query)
 	{
-		$query->where('is_admin', true)->latest();
+		$query->whereHas('permissions', function($q) {
+			$q->where(function($q) {
+				$q->where('action', 'admin_login')
+				  ->orWhere('action', Permission::WILDCARD);
+			});
+		})->latest();
 	}
 
 	public function getLabel($instance)
@@ -41,8 +48,25 @@ class UserAdminDecorator extends ModelAdminDecorator
 			new TextField('first_name'),
 			new TextField('last_name'),
 			new EmailField('email'),
-			$instance->exists ? null : new PasswordField('password'),
-			new HiddenField('is_admin', true)
+			$this->getPasswordFieldForUser($instance)
 		]);
+	}
+
+	protected function getPasswordFieldForUser($user)
+	{
+		if ( ! $user->exists) {
+			$password = new PasswordField('password');
+		} elseif ($this->isUserCurrentAuthenticatedUser($user)) {
+			$password = new PasswordField('password', ['label' => 'Change Password']);
+		} else {
+			$password = null;
+		}
+
+		return $password;
+	}
+
+	protected function isUserCurrentAuthenticatedUser($user)
+	{
+		return Auth::id() === $user->id;
 	}
 }