+++ src/Http/Controllers/AdminController.php
@@ -3,6 +3,7 @@
 use Bozboz\Admin\Models\User;
 use Illuminate\Routing\Controller;
 use Redirect, Auth, View, Input;
+use Bozboz\Permissions\Facades\Gate;
 
 class AdminController extends Controller
 {
@@ -21,11 +22,11 @@ class AdminController extends Controller
 	public function postLogin()
 	{
 		$input = Input::only('email', 'password');
-		$input['is_admin'] = true;
 
-		if (Auth::attempt($input)) {
+		if (Auth::attempt($input) && Gate::allows('admin_login')) {
 			return Redirect::intended('admin');
 		} else {
+			Auth::logout();
 			return Redirect::back()->withInput()->with('error', true);
 		}
 	}